keeps and vaults work

+++ Services/VaultsService.cs
@@ -13,9 +13,11 @@ namespace Keepr.Services
     {
       _repo = repo;
     }
-    internal IEnumerable<Vault> Get()
+    internal IEnumerable<Vault> GetByUserId(string userId)
     {
-      return _repo.Get();
+      var exists = _repo.GetByUserId(userId);
+      if (exists == null) { throw new Exception("Invalid ID"); }
+      return exists;
     }
     internal Vault GetById(int id)
     {
@@ -25,13 +27,16 @@ namespace Keepr.Services
     }
     internal Vault Create(Vault newVault)
     {
-      _repo.Create(newVault);
-      return newVault;
+      return _repo.Create(newVault);
     }
-    internal string Delete(int id)
+    internal string Delete(string userId, int id)
     {
       var exists = _repo.GetById(id);
       if (exists == null) { throw new Exception("Invalid Id"); }
+      if (exists.UserId != userId)
+      {
+        throw new Exception("I can't let you do that");
+      }
       _repo.Delete(id);
       return "Successfully Deleted";
     }
